+++ cmd/commands.go
@@ -1,0 +1,5 @@
+package cmd
+
+func Execute() {
+
+}
