--- grpc/proto/service.proto
@@ -1,34 +1,0 @@
-syntax = "proto3";
-
-package proto;
-
-option go_package =".";
-
-
-message Request {
-  int64 a = 1;
-  int64 b = 2;
-}
-
-message Response {
-  int64 result = 1;
-}
-
-message JobRequest {
-  string fileContent = 1;
-  string cronCommand = 2;
-  string fileName = 3;
-}
-
-message JobResponse {
-  string body = 1;
-}
-
-service AddService {
-  rpc Add(Request) returns (Response);
-  rpc Multiply(Request) returns (Response);
-}
-
-service JobService {
-  rpc ScheduleJob(JobRequest) returns (JobResponse);
-}
